Data.Icao.Lang

Description

Core data types and parsers.

Documentation

data FreeText

Free text, containing any character but /, - and parentheses.

endOfFieldParser :: Parser ()

A field is terminated - followed by some free text that needs to be further parsed. If the next character is not a - this signifies the end of the message and the main parser will expected a ')'. This is specially usefull when the parsed field may or may not be a terminal field. TODO: consider using this for all fields.

freeTextParser :: Parser FreeText

FreeText parser

mkFreeText :: MonadFail m => String -> m FreeText

FreeText smart constructor. Fails if given string is not a valid free text.