Data.Icao.Location

Description

Provides data types and functions pertaining to locations in accordance with the ICAO 4444 edition 2016 standard.

Data

data Aerodrome

the name of an aerodrome, 4 uppercase characters.

data Bearing

Bearing in degrees magnetic, in range [0 .. 359].

data Distance

Distance in nautical miles, always positive.

data Latitude

Latitude in decimal degrees, in range [-90 .. 90].

data Longitude

Longitude in decimal degrees, in range [-180 .. 180].

data PointNameCode

5 letters pronounceable 'name-code' (5LNC) to designate a Significant Point.

data GeographicPosition

Position expressed by the decimal latitude and longitude.

Constructors

GeographicPosition

Fields

latitude :: Latitude

longitude :: Longitude

data BearingDistancePosition

Position expressed as a bearing and distance from a reference point.

Constructors

BearingDistancePosition

Fields

reference :: PointNameCode

'name-code' of the reference point.

bearing :: Bearing

Bearing from the reference point

distance :: Distance

Distance from the reference point

data SignificantPoint

a significant point

Constructors

CodedDesignator PointNameCode
Point coded designator (e.g. RASMU), 2 to 5 characters.

Position GeographicPosition
Point Position; latitude and longitude in decimal degrees.

BearingDistance BearingDistancePosition
Bearing and distance from a reference point, bearing in degrees, distance in nautical miles.

Parsers

aerodromeParser :: Parser Aerodrome

Aerodrome Parser.

significantPointParser :: Parser SignificantPoint

SignificantPoint Parser.

parseAerodrome :: String -> Either Error Aerodrome

Parses the given textual representation of an Aerodrome. return either an Error (Left) or the parsed Aerodrome (Right).

parseSignificantPoint :: String -> Either Error SignificantPoint

Parses the given textual representation of a SignificantPoint. return either an Error (Left) or the parsed SignificantPoint (Right).

Smart constructors

mkAerodrome :: MonadFail m => String -> m Aerodrome

Aerodrome smart constructor. Fails if the given name is not a valid.

mkCodedDesignator :: MonadFail m => String -> m SignificantPoint

CodedDesignator SignificantPoint smart constructor. Fails if the given name is not a valid.

mkPosition :: MonadFail m => Float -> Float -> m SignificantPoint

Position SignificantPoint smart constructor. Fails if the given latitude and/or longitude are not valid.

mkBearingDistance :: MonadFail m => String -> Int -> Int -> m SignificantPoint

BearingDistance SignificantPoint smart constructor. Fails if the given name andor bearing andor distance are not a valid.