Data.Icao.SupplementaryInformation

Description

ICAO Field Type 19 - Supplementary information. This field is a collection of switches all optional, use the withXXXX functions to build a new instance starting from emptySupplementaryInformation.

Data

data PersonsOnBoard

Persons on board the aircraft.

data Transmitter

Transmitter.

Constructors

UHF
frequency 243.0 MHz (UHF).

VHF
frequency 121.5 MHz (VHF).

ELT
locator transmitter (ELT).

data SurvivalEquipment

Survival equipment.

Constructors

Polar
polar survival equipment.

Desert
desert survival equipment.

Maritime
maritime survival equipment.

Jungle
jungle survival equipment.

data LifeJacket

Life Jacket.

Constructors

WithLight
life jacket equipped with lights.

WithFluorescein
life jacke equipped with fluorescein.

WithRadioUHF
life jacket radio equipped with UHF on frequency 243.0 MHz.

WithRadioVHF
life jacket radio is equipped with VHF on frequency 121.5 MHz.

data Dinghies

Dinghies.

data SupplementaryInformation

Supplementary information data.

Constructors

SupplementaryInformation

Fields

fuelEndurance :: Maybe Hhmm

the fuel endurance in hours and minutes.

personsOnBoard :: Maybe PersonsOnBoard

the total number of persons on board, when so prescribed by the appropriate ATS authority.

availableTransmitters :: [Transmitter]

special transmitter(s).

survivalEquipments :: [SurvivalEquipment]

survival equipment(s) carried on board.

lifeJackets :: [LifeJacket]

type of life jackets carried on board.

dinghies :: Dinghies

description of the dinghies caried on board.

aircraftDescription :: Maybe FreeText

the colour of the aircraft and any Significant markings (this may include the aircraft registration).

otherRemarks :: Maybe FreeText

plain language indicating any other survival equipment carried and any other useful remarks.

pilotInCommand :: Maybe FreeText

the name of the pilot-in-command and possibly the contact phone.

Builders

emptySupplementaryInformation :: SupplementaryInformation

Returns empty SupplementaryInformation.

withFuelEndurance :: Hhmm -> SupplementaryInformation -> SupplementaryInformation

Sets the aircraft fuel endurance.

withPersonsOnBoard :: PersonsOnBoard -> SupplementaryInformation -> SupplementaryInformation

Sets the number of persons on board the aircraft.

withAvailableTransmitters :: [Transmitter] -> SupplementaryInformation -> SupplementaryInformation

Sets the available transmitters on board the aircraft.

withSurvivalEquipments :: [SurvivalEquipment] -> SupplementaryInformation -> SupplementaryInformation

Sets the available survival equipements on board the aircraft.

withAircraftDescription :: FreeText -> SupplementaryInformation -> SupplementaryInformation

Sets the aircraft description.

withLifeJackets :: [LifeJacket] -> SupplementaryInformation -> SupplementaryInformation

Sets the details of the life jackets available on board the aircraft.

withDinghies :: Dinghies -> SupplementaryInformation -> SupplementaryInformation

Sets the dinghies.

withOtherRemarks :: FreeText -> SupplementaryInformation -> SupplementaryInformation

Sets the othe remarks.

withPilotInCommand :: FreeText -> SupplementaryInformation -> SupplementaryInformation

Sets the detail of the pilot in command.

Smart constructors

mkPersonsOnBoard :: MonadFail m => Int -> m PersonsOnBoard

PersonsOnBoard smart constructor. Fails if given number is not in range [1 .. 999].

mkDinghies :: MonadFail m => Maybe Int -> Maybe Int -> Bool -> Maybe String -> m Dinghies

Dinghies smart constructor. Fails if dinghies details are invalid.