Data.Icao.Time

Description

Provides data types and functions pertaining to time, duration and dates in accordance with the ICAO 4444 edition 2016 standard.

Data

data Hhmm

Hhmm represents a time or duration expressed with hours and minutes only in the range 0000 .. 2359.

data DayTime

day and time.

data Date

Date represents a date expressed zith year, month and day.

Parsers

hhmmParser :: Parser Hhmm

Hhmm parser.

dateParser :: Parser Date

Date parser.

dayTimeParser :: Parser DayTime

DayTime parser.

parseHhmm :: String -> Either Error Hhmm

Parses the given textual representation of a Hhmm. return either an Error (Left) or the parsed Hhmm (Right).

parseDate :: String -> Either Error Date

Parses the given textual representation of a Date. return either an Error (Left) or the parsed Date (Right).

parseDayTime :: String -> Either Error DayTime

Parses the given textual representation of a DayTime. return either an Error (Left) or the parsed DayTime (Right).

Smart constructors

mkHhmm :: MonadFail m => Int -> Int -> m Hhmm

Hhmm smart constructor. Fails if given hour and/or minute are not valid.

mkDate :: MonadFail m => Int -> Int -> Int -> m Date

Date smart constructor. Fails if given year andor month andor day are not valid.

mkDayTime :: MonadFail m => Int -> Int -> Int -> m DayTime

DayTime smart constructor. Fails if given day andor hour andor minute are not valid.